requirements.txt
sentence-transformers
torch
pypdf
scikit-learn
numpy
pandas
rich
pyyaml
keybert
